Merge dev to main with New Release (CLI and GUI)

+++ main.py
@@ -1,44 +1,42 @@
-from os import makedirs, listdir
-from os.path import join, exists
-from pathlib import Path
-from time import sleep, time
-from tqdm import tqdm
-import requests
 import json
 from configparser import ConfigParser
+from os import makedirs, listdir
+from os.path import join, exists
+from time import sleep
+
+import requests
 from seedr_client import SeedrHandler
-from datetime import datetime
+from tqdm import tqdm
+
+from auto_seedr.utils import logging
 
 
-# import atexit
+def setup(config_file='config.ini'):
+    __config = ConfigParser()
+    __username = input("Enter your seedr email: ")
+    __password_ = input("Enter your seedr password: ")
+    _torrent_directory = input("Enter your Torrent folder default(torrents): ")
+    _download_directory = input("Enter your download folder default(downloads): ")
+    _chunkiness = input("Enter your chunk size (1024, 8192) default(1024): ")
 
+    if not _chunkiness:
+        _chunkiness = '1024'
 
-def setup():
-    username = input("Enter your seedr email: ")
-    password_ = input("Enter your seedr password: ")
-    torrent_directory = input("Enter your Torrent folder: ")
-    # download_directory = input("Enter your download folder: ")
-    chunkiness = input("Enter your chunk size (1024, 8192) put 1024 to default:  ")
+    if not _torrent_directory:
+        _torrent_directory = 'torrents'
 
-    config['seedr'] = {'user': username,
-                       'password': password_,
-                       'folder': torrent_directory,
-                       'chunk_size': chunkiness}
-    # 'download_folder': download_directory
+    if not _download_directory:
+        _download_directory = 'downloads'
 
-    with open('cred.ini', 'w') as configfile:
-        config.write(configfile)
+    __config['SEEDR'] = {'user': __username, 'password': __password_}
+
+    __config['APP'] = {'torrent_folder': _torrent_directory, 'chunk_size': _chunkiness,
+                       'download_folder': _download_directory}
+
+    with open(config_file, 'w') as configfile:
+        __config.write(configfile)
+
     print("Setup complete")
-
-
-def datetime_to_timestamp(mode):
-    stamp = datetime.now()
-    if mode == 1:
-        return stamp  # use in the log.txt for logging eg:2022-12-27 10:09:20.430322
-    elif mode == 2:
-        return stamp.strftime("%Y-%m-%d %H:%M:%S")  # use as a file name and folder name eg:2022-12-27 10:09:20
-    else:
-        raise ValueError("mode must be 1 or 2")
 
 
 def get_progression_data(url):
@@ -61,9 +59,8 @@ def is_folder_ok(folder_name, folder_id=None):
         for _folder in data['folders']:
             if _folder['folder_name'] == _folder_name:
                 return _folder['folder_id']
-        else:
-            print('Error not found folder id')
-            return None
+        print('Error not found folder id')
+        return None
 
     folder_ok = False
     if not folder_id:
@@ -79,7 +76,7 @@ def is_folder_ok(folder_name, folder_id=None):
 
 
 def upload_torrent(filename):
-    progression_data = seedr.add_torrent(torrent=f'torrent/{filename}', folder_id=-1, check_size=True)
+    progression_data = seedr.add_torrent(torrent=join(torrent_folder, filename), folder_id=-1, check_size=True)
 
     # torrent_id = progression_data.get('torrent_id')
     file_name = progression_data.get('file_name')
@@ -91,20 +88,19 @@ def upload_torrent(filename):
     return file_name, is_folder_ok(file_name)
 
 
-# remove all the progress bar to make it faster and more efficient
+# Remove all the progress bar to make it faster and more efficient
 def fast_download(url, path, file):
     print(f"\033[92m Downloading {file} to {path} \033[0m")
     if not exists(path):
         makedirs(path)
     with requests.get(url, stream=True) as r:
         r.raise_for_status()
-        # total_size = int(r.headers.get('Content-Length', 0))
         with open(join(path, file), 'wb') as f:
             for chunk in r.iter_content(chunk_size=1024):
                 f.write(chunk)
 
 
-def download_torrent(folder_id, root_dir='download'):
+def download_torrent(folder_id):
     def download_file(url, path, file):
         print(f"\033[92m Downloading {file} to {path} \033[0m")
         if not exists(path):
@@ -124,73 +120,50 @@ def download_torrent(folder_id, root_dir='download'):
         _data = seedr.get_folder(_folder)
         file_to_download += _data["files"]
         folders_to_look += [fol["folder_id"] for fol in _data["folders"]]
-    start_time = time()
-    print(start_time)
+
     for _file in file_to_download:
         download_file(seedr.get_file(_file.get('folder_file_id'))['download_url'],
-                      join(root_dir, _file.get('folder_path')), _file.get('file_name'))
-    end_time = time()
-
-    folder_name = seedr.get_folder(folder_id)["folder_name"]
-    root_directory = Path(join(root_dir, folder_name))
-    with open('speed.txt', 'a') as f:
-        f.write(
-            f"{end_time - start_time}, {folder_name}, {sum(f.stat().st_size for f in root_directory.glob('**/*') if f.is_file())} \n")
+                      join(download_folder, _file.get('folder_path')), _file.get('file_name'))
 
 
 def delete_folder(parent_folder_id):
     seedr.delete_folder(parent_folder_id)
 
 
-def logging(faulty_torrents_):
-    with open('log.txt', 'a') as f:
-        for i in faulty_torrents_:
-            f.write(f"{datetime_to_timestamp(1)}: {i} \n")
-
-
-# This function for logging the total running time of the program
-def trigger_while_exit():
-    end_time = time()
-    with open('time_log.txt', 'a') as f:
-        f.write(f"Program started {datetime_to_timestamp(1)}\tProgram runtime: {end_time - program_start_time}\n")
-
-
 def directory_download():
     if not exists(torrent_folder):
         makedirs(torrent_folder)
-    file_list = listdir(torrent_folder)
-    for i in file_list:
-        try:
-            folder_name, folder_id = upload_torrent(i)
-            download_torrent(folder_id)
-            delete_folder(folder_id)
-        except Exception as e:
-            faulty_torrents.append(f'{i} - {e}')
-            pass
-    logging(faulty_torrents)
+    file_list = [f for f in listdir(torrent_folder) if f.endswith(".torrent")]
+    if file_list:
+        for i in file_list:
+            try:
+                _, folder_id = upload_torrent(i)
+                download_torrent(folder_id)
+                delete_folder(folder_id)
+            except Exception as e:
+                logging(f'{i} - {e}')
+    else:
+        print(f"No torrent files found in {torrent_folder}")
 
 
-if not exists('cred.ini'):
-    setup()
-else:
+if __name__ == '__main__':
+    if not exists('config.ini'):
+        setup()
+
     config = ConfigParser()
-    config.read('cred.ini')
-    user = config['seedr']['user']
-    password = config['seedr']['password']
-    torrent_folder = config['seedr']['torrent_folder']
-    download_folder = config['seedr']['download_folder']
-    chunk_size = config['seedr']['chunk_size']
+    config.read('config.ini')
+
+    user = config['SEEDR']['user']
+    password = config['SEEDR']['password']
+    torrent_folder = config['APP']['torrent_folder']
+    download_folder = config['APP']['download_folder']
+    chunk_size = config['APP']['chunk_size']
 
     seedr = SeedrHandler(email=user, password=password)
 
-    faulty_torrents = []
-    program_start_time = time()
-
     directory_download()
 
-# atexit.register(trigger_while_exit)
 # TODO: add a function to check if the file is already downloaded
 # TODO: add argument parser using argparse (fastdownload, progressbar download,)
-# TODO: Multiple torrent download from multiple seedr accounts to use maximum bandwidth
-#  from isp and avoid limits of seedr server bandwidth
+# TODO: Multiple torrent download from multiple seedr accounts to use maximum bandwidth from isp and avoid limits of seedr server bandwidth
 # FIXME: fix the progress bar
